feat: add client.disconnect event Disconnect events will be emitted to relevant client ports in three cases: - A connection is removed. All ports for the given origin are disconnected - Wallet is locked. All ports for all origins are disconnected - Content script looses connectivity to the background
The last case is the consequence of the first two, ie. the background
script never technically sends a disconnect notification, but simply
closes ports. However non-persistent background scripts and service workers
in MV3 extensions may be suspended by the browser. This is not observable
from service workers (ie the background type we use), so we instead have
to observe it form the content script.