feat: Add multi-network
dApps should now specify a `chainId` when calling
`client.connect_wallet` instead of relying on the extension
having a only one configure network. This also bleeds into
the approval flow for new connections now passing the requested
chainId and asking the UI to return a valid networkId.

`client.get_chain_id` will work for now with either the chain ID
of the default network for the extension, or the chain ID that
the dApp requested when connecting.

Some helper methods were added to both the connections collection
and the networks collection to aid in picking the right network
configuration based on either a network ID or a chain ID. When
asking from a chain ID, the first network config found will be
returned. The order is non-deterministic.