fix: connection access time and disconnect
`client.disconnect_wallet` is now a no-op. We may change the behaviour
in the future to allow dApps to remove themselves as connections,
however this would break with how this API is implemented in
CLI and Desktop.

The accessedAt time for a connection is now updated each time the user
initially loads a dApp, given that we introduced auto-connect.